# openapi.yaml
openapi: 3.1.0
info:
  title: Event Hunter API
  version: 0.1.0
  summary: 事件驅動架構治理與業務事故鑑識 API 草案
  description: |
    Event Hunter 是建立在 Kafka、ClickHouse 與 OpenTelemetry 資料之上的
    業務級事件鑑識平台。本文件保留完整的未來 API 草案；MVP 的 Event Hunter
    使用者介面聚焦 Business Timeline、Investigation、固定 Domain Pattern 與
    Evidence Bundle，另包含 Grafana 業務告警接入、事件品質聚合與可展示的
    Outbox-to-Timeline 管線等 supporting capabilities。
    Event Catalog、Topic Registry、Event Hunter 自建 Runtime Quality UI 與 Replay
    是延後功能，先由 Schema Registry／Kafka tooling／Grafana OSS 等既有元件處理。
    Grafana OSS 負責 Logs／Metrics／Traces、Dashboard、Explore、Correlations 與
    Alerting；Event Hunter 只接收需要業務調查的 Grafana Alert Webhook，並保存參照。

    ### 重要邊界

    - 第一版 Pattern Engine 只做固定規則的唯讀調查，不直接修改正式業務資料。
    - `forensics_events` 等事件資料位於 ClickHouse；案件、權限與設定位於 PostgreSQL。
    - Temporal 是預設關閉的選用能力，只負責低頻、可重試、可能等待人工核准的案件流程；MVP 核心功能不依賴 Temporal。
    - Kafka Consumer 的事件重試由 Consumer Framework／Retry Topic／DLQ 處理；Event Hunter 只有在來源提供 attempt 與錯誤資訊時才呈現重試摘要。
    - Production Redrive（把歷史事件重新送回正式事件流）不在本 API 草案範圍內。
  contact:
    name: Event Hunter Team
  license:
    name: Proprietary / Internal Draft
servers:
  - url: http://localhost:28333
    description: 本機開發環境
  - url: https://event-hunter.example.com
    description: 預留的部署環境（尚未實作）
...
paths:
...
  /api/v1/event-checks/evaluations:
    x-mvp-status: implemented-event-check-api
    x-owner: event-check
    post:
      tags: [Event Check]
      operationId: evaluateEventCheck
      summary: 即時解析事件範圍並執行 Check Model
      description: >-
        Stateless query-time evaluation。NO_DATA、IDENTIFIER_SELECTION_REQUIRED、
        MODEL_SELECTION_REQUIRED 與 NO_APPLICABLE_MODEL 都以 200 回傳可操作的產品狀態；
        本 operation 不建立 PostgreSQL Snapshot，也不接受 SQL、任意 URL 或 production command。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './contracts/event-check/event-check-evaluation.schema.json#/$defs/EvaluationRequest'
      responses:
        '200':
          description: 事件範圍解析完成；可能已評估、需要選擇或沒有資料／Model。
          content:
            application/json:
              schema:
                $ref: './contracts/event-check/event-check-evaluation.schema.json'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: 尚未建立有效 Session。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
        '504':
          $ref: '#/components/responses/DependencyTimeout'
  /api/v1/check-models:
    x-mvp-status: implemented-event-check-api
    x-owner: event-check
    get:
      tags: [Check Models]
      operationId: listCheckModels
      summary: 列出 immutable Check Model Registry
      description: 回傳已發布 Flow Models 與 Global Checks；本階段不提供 runtime CRUD 或發布 API。
      responses:
        '200':
          description: Check Model catalog。
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: './contracts/check-models/check-model-registry.schema.json#/$defs/RegistryEntry'
        '401':
          description: 尚未建立有效 Session。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /api/v1/check-snapshots:
    x-mvp-status: implemented-event-check-api
    x-owner: event-check
    get:
      tags: [Event Check]
      operationId: listCheckSnapshots
      summary: 列出已保存的 immutable Event Check 結果
      description: >-
        Saved Results read-model。以建立時間與 UUID 進行穩定 keyset 分頁，可依識別碼或
        Check status 篩選；僅回傳摘要、固定 evaluation request 與案件連結數，不包含 event payload。
      parameters:
        - name: identifier
          in: query
          description: 依保存時的 identifier value 搜尋，最多 200 字元。
          schema:
            type: string
            maxLength: 200
        - name: check_status
          in: query
          schema:
            type: string
            enum: [NO_DATA, IN_PROGRESS, CONFORMANT, DEVIATED, INCONCLUSIVE, AMBIGUOUS]
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: cursor
          in: query
          description: 上一頁回應的 `next_cursor`，第一次查詢不傳。
          schema:
            type: string
      responses:
        '200':
          description: Saved Results page。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckSnapshotPage'
        '401':
          description: 尚未建立有效 Session。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
    post:
      tags: [Event Check]
      operationId: createCheckSnapshot
      summary: 驗證 evaluation hashes 並保存 immutable Snapshot
      description: >-
        只允許 INVESTIGATOR／ADMIN。伺服器使用相同 as_of 與 pinned Model version 重算；
        event set 或 result 改變時回 409 EVALUATION_CHANGED，不接受前端 result JSON 作為正式資料。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './contracts/event-check/check-snapshot.schema.json#/$defs/CreateCheckSnapshotRequest'
      responses:
        '200':
          description: 相同 actor 與 Idempotency-Key 已成功，回傳原 Snapshot。
          content:
            application/json:
              schema:
                $ref: './contracts/event-check/check-snapshot.schema.json'
        '201':
          description: Snapshot、event refs、relations、findings 與 Audit 已 transactionally 建立。
          content:
            application/json:
              schema:
                $ref: './contracts/event-check/check-snapshot.schema.json'
        '401':
          description: 尚未建立有效 Session。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: >-
            `EVALUATION_CHANGED`、`MODEL_VERSION_UNAVAILABLE` 或 `IDEMPOTENCY_KEY_REUSED`。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventCheckConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
        '504':
          $ref: '#/components/responses/DependencyTimeout'
  /api/v1/check-snapshots/{snapshotId}:
    x-mvp-status: implemented-event-check-api
    x-owner: event-check
    get:
      tags: [Event Check]
      operationId: getCheckSnapshot
      summary: 取得 immutable Check Snapshot
      parameters:
        - $ref: '#/components/parameters/CheckSnapshotId'
      responses:
        '200':
          description: >-
            Snapshot 與保存當下的 event reference 可用狀態。ClickHouse TTL／Evidence Archive 的讀取時
            source revalidation 由選用的 EH-ECM-007 交付。
          content:
            application/json:
              schema:
                $ref: './contracts/event-check/check-snapshot.schema.json'
        '401':
          description: 尚未建立有效 Session。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/check-findings/{findingId}/feedback:
    x-mvp-status: implemented-event-check-api
    x-owner: investigation
    parameters:
      - $ref: '#/components/parameters/CheckFindingId'
    patch:
      tags: [Investigation, Event Check]
      operationId: classifyCheckFinding
      summary: 更新 generalized Check Finding 的人工判定
      description: >-
        Finding 本身 append-only；只更新 feedback current state。相容期舊 classifyPatternFinding
        透過 adapter 操作同一筆 feedback，不建立第二套狀態。
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClassifyCheckFindingRequest'
      responses:
        '200':
          description: Feedback 與 Audit 已在同一 transaction 更新。
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckFindingFeedback'
        '401':
          description: 尚未建立有效 Session。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    demoSession:
      type: apiKey
      in: cookie
      name: eh_demo_session
      description: MVP 示範用的簽署 HttpOnly Cookie；不對應使用者資料表。
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 正式環境預留，由外部 OIDC／Identity Provider 核發；MVP 使用 `demoSession`。
    grafanaHmac:
      type: apiKey
      in: header
      name: X-Grafana-Alerting-Signature
      description: |
        Grafana Webhook Contact Point 產生的 hex HMAC-SHA256。伺服器必須搭配
        `X-Grafana-Alerting-Timestamp` 與 raw request body 驗證，使用 constant-time comparison。
  parameters:
...
    CheckSnapshotId:
      name: snapshotId
      in: path
      required: true
      description: Immutable Check Snapshot UUID。
      schema:
        type: string
        format: uuid
    CheckFindingId:
      name: findingId
      in: path
      required: true
      description: Persisted generalized Check Finding UUID。
      schema:
        type: string
        format: uuid
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: GET 回應的 ETag，例如 `W/"investigation-123-lock-v7"`。版本不符時回傳 `409 Conflict`。
      schema:
        type: string
        minLength: 1
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Actor-scoped Snapshot create retry key；相同 key 搭配不同 request 回 409。
      schema:
        type: string
        minLength: 8
        maxLength: 200
  headers:
    ETag:
      description: 資源版本，用於下一次更新的 `If-Match` 樂觀鎖。
      schema:
        type: string
        example: W/"investigation-123-lock-v7"
    RateLimitLimit:
      description: 目前固定窗口允許的最大請求數。
      schema:
        type: integer
    RateLimitRemaining:
      description: 目前固定窗口剩餘的請求數。
      schema:
        type: integer
    RateLimitReset:
      description: 距離目前窗口重設的秒數。
      schema:
        type: integer
    RetryAfter:
      description: 被限流後可重試前應等待的秒數。
      schema:
        type: integer
  responses:
    BadRequest:
      description: 請求格式或查詢條件錯誤。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Request Schema 驗證失敗。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    NotFound:
      description: 找不到指定資源。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: 目前角色缺少執行此操作所需的權限。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: 樂觀鎖版本衝突或目前狀態不允許此操作。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ConflictError'
    RateLimited:
      description: 同一遠端位址在固定窗口內超過 API 請求上限。
      headers:
        RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    DependencyUnavailable:
      description: 必要的 PostgreSQL 或 ClickHouse 來源無法使用。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    DependencyTimeout:
      description: 必要來源超過伺服器端查詢期限。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
          example: RESOURCE_NOT_FOUND
        message:
          type: string
          description: 給使用者或工程師看的錯誤說明。
          example: 找不到指定的調查案件。
        request_id:
          type: string
          description: 供客服與日誌查找的請求 ID。
          example: req_01J5QZ7F9P
        details:
          type: object
          additionalProperties: true
    ValidationError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            field_errors:
              type: array
              items:
                type: object
                required: [field, message]
                properties:
                  field:
                    type: string
                  message:
                    type: string
    ConflictError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          required: [current_lock_version]
          properties:
            current_lock_version:
              type: integer
              format: int64
              description: 伺服器目前版本；客戶端應重新 GET 後再提交。
            current_resource:
              type: object
              additionalProperties: true
    EventCheckConflictError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          required: [code]
          properties:
            code:
              type: string
              enum: [EVALUATION_CHANGED, MODEL_VERSION_UNAVAILABLE, IDEMPOTENCY_KEY_REUSED]
            current_event_set_hash:
              type: [string, 'null']
              pattern: '^[0-9a-f]{64}$'
            current_evaluation_hash:
              type: [string, 'null']
              pattern: '^[0-9a-f]{64}$'
          description: >-
            EVALUATION_CHANGED 時回傳目前 hashes，要求使用者重新檢視；伺服器不得靜默保存不同結果。
    CheckSnapshotSummary:
      type: object
      additionalProperties: false
      required: [id, created_by, created_by_role, created_at, evaluation_request, as_of, source_health_status, model, check_status, event_count, finding_count, linked_case_count]
      properties:
        id:
          type: string
          format: uuid
        created_by:
          type: string
        created_by_role:
          type: string
          enum: [INVESTIGATOR, ADMIN]
        created_at:
          type: string
          format: date-time
        evaluation_request:
          $ref: './contracts/event-check/event-check-evaluation.schema.json#/$defs/EvaluationRequest'
        as_of:
          type: string
          format: date-time
        source_health_status:
          type: string
          enum: [HEALTHY, STALE, PARTIAL, UNAVAILABLE]
        model:
          type: object
          additionalProperties: false
          required: [id, version, kind]
          properties:
            id:
              type: string
            version:
              type: integer
              minimum: 1
            kind:
              type: string
              enum: [FLOW, GLOBAL_CHECK]
        check_status:
          type: string
          enum: [NO_DATA, IN_PROGRESS, CONFORMANT, DEVIATED, INCONCLUSIVE, AMBIGUOUS]
        event_count:
          type: integer
          minimum: 0
        finding_count:
          type: integer
          minimum: 0
        linked_case_count:
          type: integer
          minimum: 0
    CheckSnapshotPage:
      type: object
      additionalProperties: false
      required: [items, page_size, next_cursor]
      properties:
        items:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/CheckSnapshotSummary'
        page_size:
          type: integer
          minimum: 1
          maximum: 100
        next_cursor:
          type: [string, 'null']
...
    ClassifyPatternFindingRequest:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          type: string
          enum: [CONFIRMED, FALSE_POSITIVE, NEEDS_REVIEW]
    PatternFindingFeedback:
      type: object
      additionalProperties: false
      required: [finding_id, status, actor_id, actor_role, updated_at, lock_version]
      properties:
        finding_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [UNREVIEWED, CONFIRMED, FALSE_POSITIVE, NEEDS_REVIEW]
        actor_id:
          type: string
          description: UNREVIEWED 時為空字串。
        actor_role:
          type: string
          description: UNREVIEWED 時為空字串。
        updated_at:
          type: [string, 'null']
          format: date-time
        lock_version:
          type: integer
          format: int64
          minimum: 0
    ClassifyCheckFindingRequest:
      $ref: '#/components/schemas/ClassifyPatternFindingRequest'
    CheckFindingFeedback:
      $ref: '#/components/schemas/PatternFindingFeedback'
...
security:
  - demoSession: []
  - bearerAuth: []